requirements.txt
Django==1.9.4
django-activity-stream==0.6.1
django-allauth==0.25.2
django-braces==1.8.1
django-cors-headers==1.1.0
django-countries==3.4.1
django-filter==0.13.0
django-oauth-toolkit==0.10.0
django-rest-auth==0.7.0
django-rest-swagger==0.3.5
django-tagulous==0.11.1
djangorestframework==3.3.3
djangorestframework-jwt==1.7.2
dry-rest-permissions==0.1.6
freeze==1.0.9
gunicorn==19.4.5
mysqlclient==1.3.7
oauthlib==1.0.3
Pillow==3.1.1
PyJWT==1.4.0
python-openid==2.2.5
PyYAML==3.11
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
wheel==0.24.0
